==> 1. loading_TEMU.ipynb <==
## LOADING LIBRARIES

import pandas as pd

## Loading Temu Kaggle File 

df_temu_kaggle = pd.read_csv("temu_product_sales_dataset.csv")
df_temu_kaggle.leve_1_category_name.value_counts()

## Separating or creating csv file for the selected category of Temu

#filtering category, choosing necessary columns, dropping null & duplicates & index, resetting index and choosing top 100 products on basis of sold_units
df_home = df_temu_kaggle[df_temu_kaggle["leve_1_category_name"].isin(["Toys & Games"])][["leve_1_category_name", "title", "price", "sales_info", "comment_num"]]
df = df_home.drop_duplicates().dropna(how='any').reset_index().head(100).drop(["index"],axis=1)


#changing data types and as well as necessary steps
df["sales_info"] = df["sales_info"].str.replace(" sold", "").str.replace(',', '').astype(int)
df["comment_num"] = df["comment_num"].astype(int)
df["leve_1_category_name"] = df["leve_1_category_name"].str.replace("Toys & Games", "5") #assigning categorey with an id


# renaming columns
df = df.rename(columns={"leve_1_category_name": "category_id", "sales_info": "n_sold_units", "comment_num":"n_reviews"})

#assigning store id
df["store_id"] = 3

#re-ordering the columns
df = df[['store_id', 'category_id', 'title', 'price', 'n_reviews', 'n_sold_units']]

df.to_csv("temu_TOYS&Games.csv")
df
