==> 3. webscrapping_AMAZON.ipynb <==
import pandas as pd
import requests
import time
import random

from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.edge.options import Options
from bs4 import BeautifulSoup

## Webscrapping Aamazon.de with selenium

total_pages = 2
driver = webdriver.Edge()
allproducts = []
my_dict = {}
titles = []
prices = []
reviews = []


for page in range(1, total_pages + 1):    
    url = f"https://www.amazon.de/-/en/gp/bestsellers/ce-de/ref=zg_bs_pg_{page}_ce-de?ie=UTF8&pg={page}"
    driver.get(url)
    driver.maximize_window()

    body = driver.find_element(By.TAG_NAME, "body")
    for i in range(10):
        body.send_keys(Keys.PAGE_DOWN)
        time_2wait = random.uniform(2, 3) * 2
        time.sleep(time_2wait)
    
    products_pg = driver.find_elements(By.ID, "gridItemRoot")



    for prod in products_pg:
        try:
            title = prod.find_element(By.CLASS_NAME, "_cDEzb_p13n-sc-css-line-clamp-3_g3dy1").text
            titles.append(title)
        except:
            titles.append("no title")
        try:
            price = prod.find_element(By.CLASS_NAME, "_cDEzb_p13n-sc-price_3mJ9Z").text
            prices.append(price)
        except:
            prices.append(0)
        try:
            n_review = prod.find_element(By.CLASS_NAME, "a-size-small").text
            reviews.append(n_review)
        except:
            reviews.append(0)


my_dict = {"title": titles,
           "price": prices,
           "review": reviews}


df = pd.DataFrame.from_dict(my_dict, orient="columns")
df

#df.to_csv("amazon_ELECTRONICS.csv")
