==> 4. temu_MASTER.ipynb <==
import pandas as pd

## loading the CSV file for each category of TEMU and merging them

df1 = pd.read_csv("temu_TOYS&Games.csv")
df2 = pd.read_csv("temu_ELECTRONICS.csv")
df3 = pd.read_csv("temu_BEAUTY_&_HEALTH.csv")
df4 = pd.read_csv("temu_FASHION.csv")
df5 = pd.read_csv("temu_Home_&_Kitchen.csv")

df_temu = pd.concat([df1,df2,df3,df4,df5], axis=0)
df_temu_final = df_temu.drop(["Unnamed: 0"],axis=1).reset_index(drop=True)

df_temu_final.to_csv("temu_ALLDATA.csv") 

df_temu_final
