==> 5. ali_express_MASTER.ipynb <==
import pandas as pd

## loading webscrapped csv one by one and cleaning them

first toys & games

df_ali_TG = pd.read_csv("ali_express_TOYS&GAMES.csv")
df1 = df_ali_TG.drop(["Unnamed: 0"],axis=1)
df1

#cleaning & changing data types and as well as necessary steps
df1["prices"] = df1["prices"].str.replace("€", "").str.replace(',', '.').astype(float)
df1["sold"] = df1["sold"].str.replace(" sold", "").str.replace(',', '').str.replace('+', '').astype(int)


# renaming columns
df1 = df1.rename(columns={"names": "title", "prices": "price", "sold":"n_sold_units",})

#assigning store id & category _id
df1["store_id"] = 2
df1["category_id"] = 5

#re-ordering the columns
df1 = df1[['store_id', 'category_id', 'title', 'price', 'n_reviews', 'n_sold_units']]

df1

second electronics

#loading the csv
df_ali_electronics = pd.read_csv("ali_express_ELECTRONICS.csv")

#cleaning & changing data types and as well as necessary steps
df2 = df_ali_electronics.drop(["Unnamed: 0"],axis=1)
df2["prices"] = df2["prices"].str.replace("￡", "").astype(float)
df2["sold"] = df2["sold"].str.replace(" sold", "").str.replace(',', '').str.replace('+', '').str.replace('K', '000').astype(int)

# renaming columns, assigning store id & category _id, re-ordering the columns
df2 = df2.rename(columns={"names": "title", "prices": "price", "sold":"n_sold_units",})
df2["store_id"] = 2
df2["category_id"] = 3
df2 = df2[['store_id', 'category_id', 'title', 'price', 'n_reviews', 'n_sold_units']]
df2

third home / kitchen

#loading the csv
df_ali_home = pd.read_csv("ali_express_Home_&_GARDEN.csv")
df_ali_garden = pd.read_csv("ali_express_Home_&_LIGHT.csv")

#merging the datasets
df3 = pd.concat([df_ali_home,df_ali_garden], axis=0)

#cleaning & changing data types and as well as necessary steps
df3 = df3.drop(["Unnamed: 0"],axis=1).reset_index(drop=True)
df3["prices"] = df3["prices"].str.replace("￡", "").astype(float)
df3["sold"] = df3["sold"].str.replace(" sold", "").str.replace(',', '').str.replace('+', '').str.replace('K', '000').astype(int)

#renaming column, assigining store & category id, re ordering the columns
df3 = df3.rename(columns={"names": "title", "prices": "price", "sold":"n_sold_units",})
df3["store_id"] = 2
df3["category_id"] = 4
df3 = df3[['store_id', 'category_id', 'title', 'price', 'n_reviews', 'n_sold_units']]

df3

fourth clothing / apparel

#loading the csv
df_ali_shoes = pd.read_csv("ali_express_clothing&apparel_shoes.csv")
df_ali_men = pd.read_csv("ali_express_clothing&apparel_men.csv")
df_ali_women = pd.read_csv("ali_express_clothing&apparel_women.csv")

#merging the datasets
df4 = pd.concat([df_ali_shoes,df_ali_men,df_ali_women], axis=0)

#cleaning & changing data types and as well as necessary steps
df4 = df4.drop(["Unnamed: 0"],axis=1).reset_index(drop=True)
df4["prices"] = df4["prices"].str.replace("￡", "").astype(float)
df4["sold"] = df4["sold"].str.replace(" sold", "").str.replace(',', '').str.replace('+', '').str.replace('K', '000').astype(int)

#renaming column, assigining store & category id, re ordering the columns
df4 = df4.rename(columns={"names": "title", "prices": "price", "sold":"n_sold_units",})
df4["store_id"] = 2
df4["category_id"] = 2
df4 = df4[['store_id', 'category_id', 'title', 'price', 'n_reviews', 'n_sold_units']]

df4

fifth and final the beauty category

#loading the csv
df_ali_beauty = pd.read_csv("ali_express_BEAUTY_&_HEALTH.csv")

#cleaning & changing data types and as well as necessary steps
df5 = df_ali_beauty.drop(["Unnamed: 0"],axis=1)
df5["prices"] = df5["prices"].str.replace("￡", "").astype(float)
df5["sold"] = df5["sold"].str.replace(" sold", "").str.replace(',', '').str.replace('+', '').str.replace('K', '000').astype(int)

# renaming columns, assigning store id & category _id, re-ordering the columns
df5 = df5.rename(columns={"names": "title", "prices": "price", "sold":"n_sold_units",})
df5["store_id"] = 2
df5["category_id"] = 1
df5 = df5[['store_id', 'category_id', 'title', 'price', 'n_reviews', 'n_sold_units']]
df5

## The Master dataframe for ALI EXPRESS

#merging the datasets and converting to a final csv for merging with other online stores and analysis later
df_ali_final = pd.concat([df1,df2,df3,df4,df5], axis=0)
df_ali_final.reset_index(drop=True, inplace=True)

df_ali_final.to_csv("ali_express_ALLDATA.csv") 

df_ali_final
