==> 6. amazon_master.ipynb <==
import pandas as pd

# loading webscrapped csv one by one and cleaning them 
### also merging the "n_sold_units" from the csv

first the "beauty" category

#loading and merging
df_amzn_beauty = pd.read_csv("amazon_BEAUTY.csv")
df_beauty_su = pd.read_csv("Beauty - Sheet1.csv")
df1 = pd.concat([df_amzn_beauty,df_beauty_su], axis=1)

#cleaning
df1 = df1.drop(["Unnamed: 0"],axis=1)
df1["price"] = df1["price"].str.replace("€", "").astype(float)
df1["review"] = df1["review"].str.replace(',', '')
df1["n_sold_units"] = df1["n_sold_units"].str.replace(',', '').str.replace('+', '').str.replace('K', '000').astype(int)

# renaming columns, assigning store id & category _id, re-ordering the columns
df1 = df1.rename(columns={"review": "n_reviews"})
df1["store_id"] = 1
df1["category_id"] = 1
df1 = df1[['store_id', 'category_id', 'title', 'price', 'n_reviews', 'n_sold_units']]
df1

second the "clothing" category

#loading and merging
df_amzn_fashion = pd.read_csv("amazon_FASHION.csv")
df_fashion_su = pd.read_csv("Clothing - Sheet1.csv")
df2 = pd.concat([df_amzn_fashion,df_fashion_su], axis=1)

#cleaning
df2 = df2.drop(["Unnamed: 0"],axis=1)
df2["price"] = df2["price"].str.replace("€", "").astype(float)
df2["review"] = df2["review"].str.replace(',', '')
df2["n_sold_units"] = df2["n_sold_units"].str.replace(',', '').str.replace('+', '').str.replace('K', '000').astype(int)

# renaming columns, assigning store id & category _id, re-ordering the columns
df2 = df2.rename(columns={"review": "n_reviews"})
df2["store_id"] = 1
df2["category_id"] = 2
df2 = df2[['store_id', 'category_id', 'title', 'price', 'n_reviews', 'n_sold_units']]
df2

third the "electronics" category

#loading and merging
df_amzn_electronics = pd.read_csv("amazon_ELECTRONICS.csv")
df_electronics_su = pd.read_csv("Electronics - Sheet1.csv")
df3 = pd.concat([df_amzn_electronics,df_electronics_su], axis=1)

#cleaning
df3 = df3.drop(["Unnamed: 0"],axis=1)
df3["price"] = df3["price"].str.replace("€", "").astype(float)
df3["review"] = df3["review"].str.replace(',', '')
df3["n_sold_units"] = df3["n_sold_units"].str.replace(',', '').str.replace('+', '').str.replace('K', '000').astype(int)

# renaming columns, assigning store id & category _id, re-ordering the columns
df3 = df3.rename(columns={"review": "n_reviews"})
df3["store_id"] = 1
df3["category_id"] = 3
df3 = df3[['store_id', 'category_id', 'title', 'price', 'n_reviews', 'n_sold_units']]
df3

fourth the "Home & Kitchen" category

#loading and merging
df_amzn_home = pd.read_csv("amazon_HOME_&_kitchen.csv")
df_home_su = pd.read_csv("Home & Kitchen - Sheet1.csv")
df4 = pd.concat([df_amzn_home,df_home_su], axis=1)

#cleaning
df4 = df4.drop(["Unnamed: 0"],axis=1)
df4["price"] = df4["price"].str.replace("€", "").astype(float)
df4["review"] = df4["review"].str.replace(',', '')
df4["n_sold_units"] = df4["n_sold_units"].str.replace(',', '').str.replace('+', '').str.replace('K', '000').astype(int)

# renaming columns, assigning store id & category _id, re-ordering the columns
df4 = df4.rename(columns={"review": "n_reviews"})
df4["store_id"] = 1
df4["category_id"] = 4
df4 = df4[['store_id', 'category_id', 'title', 'price', 'n_reviews', 'n_sold_units']]
df4

fifth the "Toys & Games" category

#loading and merging
df_amzn_toys = pd.read_csv("amazon_TOYS_&_GAMES.csv")
df_toys_su = pd.read_csv("Toys & Games - Sheet1.csv")
df5 = pd.concat([df_amzn_toys,df_toys_su], axis=1)

#cleaning
df5 = df5.drop(["Unnamed: 0"],axis=1)
df5 = df5.drop_duplicates().dropna(how='any')
df5["price"] = df5["price"].str.replace("€", "").astype(float)
df5["review"] = df5["review"].str.replace(',', '')
df5["n_sold_units"] = df5["n_sold_units"].str.replace(',', '').str.replace('+', '').str.replace('K', '000').astype(int)

# renaming columns, assigning store id & category _id, re-ordering the columns
df5 = df5.rename(columns={"review": "n_reviews"})
df5["store_id"] = 1
df5["category_id"] = 5
df5 = df5[['store_id', 'category_id', 'title', 'price', 'n_reviews', 'n_sold_units']]
df5

## The Master dataframe for AMAZON

#merging the datasets and converting to a final csv for merging with other online stores and analysis later
df_amazon_final = pd.concat([df1,df2,df3,df4,df5], axis=0)
df_amazon_final.reset_index(drop=True, inplace=True)

df_amazon_final.to_csv("amazon_ALLDATA.csv") 

df_amazon_final

created two csv file for store names and category names with their ID's

df_store = pd.read_csv("online_stores_names&ID.csv")
df_store

df_category = pd.read_csv("category_names&ID.csv")
df_category.reset_index(drop=True)
